pyproject.toml
[project]
name = "sparsetriton"
version = "0.1.0"
description = "Triton-based 3D Sparse Convolution for high-performance deep learning"
authors = [
    {name = "박태준", email = "ptj0225@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.9"
keywords = ["sparse", "convolution", "triton", "deep learning", "gpu", "pytorch"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules"
]
dependencies = [
    "torch>=2.1",
    "triton>=2.1",
    "numpy",
    "backports.cached-property; python_version < '3.8'",
]

[project.optional-dependencies]
test = [
    "pytest",
    "pytorch-sparsetorch"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
